Aplicação de DDD e Clean Architecture a estrutura do projeto
