+++ TheatricalPlayersRefactoringKata.Tests/DomainTests/ComedyPlayTests.cs
@@ -2,9 +2,10 @@ using System;
 using System.Collections.Generic;
 using ApprovalTests;
 using ApprovalTests.Reporters;
+using TheatricalPlayersRefactoringKata.Entities;
 using Xunit;
 
-namespace TheatricalPlayersRefactoringKata.Tests;
+namespace TheatricalPlayersRefactoringKata.Tests.DomainTests;
 
 public class ComedyPlayTests
 {
