+++ TheatricalPlayersRefactoringKata.Tests/FormatterTests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
