+++ TheatricalPlayersRefactoringKata.Tests/FormatterTests/StatementPrinterTests.TestTextStatementExample.approved.txt
