+++ TheatricalPlayersRefactoringKata.Tests/FormatterTests/StatementPrinterTests.TestXmlStatementExample.approved.txt
