+++ TheatricalPlayersRefactoringKata/Entities/HistoricalPlay.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Entities;
 
 public class HistoricalPlay : Play
 {
