+++ TheatricalPlayersRefactoringKata/Entities/TragedyPlay.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Entities;
 
 public class TragedyPlay : Play
 {
@@ -9,7 +9,7 @@ public class TragedyPlay : Play
     public override decimal CalculateValue(int audience)
     {
         var amount = 0;
-        var playLines = this.Lines;
+        var playLines = Lines;
 
         if (playLines < 1000) playLines = 1000;
         if (playLines > 4000) playLines = 4000;
