--- TheatricalPlayersRefactoringKata/IStatementFormatter.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace TheatricalPlayersRefactoringKata;
-
-public interface IStatementFormatter
-{
-    string Print(Invoice invoice);
-}
