+++ TheatricalPlayersRefactoringKata/Interfaces/IStatementFormatter.cs
@@ -1,0 +1,9 @@
+using System;
+using TheatricalPlayersRefactoringKata.Entities;
+
+namespace TheatricalPlayersRefactoringKata.Interfaces;
+
+public interface IStatementFormatter
+{
+    string Print(Invoice invoice);
+}
